requirements.txt
numpy
pandas
scikit-learn
matplotlib
opencv-python
tensorflow
keras
Pillow
scipy